// frontend/src/api/generated/models/issueLinks.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SoftTrack
 * An open-source, self-hostable issue tracker inspired by Linear.
 * OpenAPI spec version: 0.1.0
 */
import type { IssueLinkRead } from './issueLinkRead';

export interface IssueLinks {
  blocks?: IssueLinkRead[];
  blocked_by?: IssueLinkRead[];
  relates_to?: IssueLinkRead[];
  duplicates?: IssueLinkRead[];
  duplicated_by?: IssueLinkRead[];
}